Moved Attributes to sepearte file
