+++ libopenage/unit/CMakeLists.txt
@@ -2,6 +2,7 @@ add_sources(libopenage
 	ability.cpp
 	action.cpp
 	attribute.cpp
+	attributes.cpp
 	command.cpp
 	producer.cpp
 	selection.cpp
